Small refactoring and first part of improvement in E2E Tests

+++ src/test/groovy/io/project/kyma/ace2e/utils/K8SClient.groovy
@@ -1,5 +1,6 @@
 package io.project.kyma.ace2e.utils
 
+import com.google.gson.internal.LinkedTreeMap
 import io.kubernetes.client.ApiClient
 import io.kubernetes.client.Configuration
 import io.kubernetes.client.apis.AppsV1Api
@@ -20,7 +21,7 @@ class K8SClient {
     static final String SERVICE_CATALOG_API_GROUP = "servicecatalog.k8s.io"
     static final String KUBELESS_API_GROUP = "kubeless.io"
     static final String ISTIO_API_GROUP = "networking.istio.io"
-    static final String V1ALPHA1_API_VERSION = "v1alpha1"
+    public static final String V1ALPHA1_API_VERSION = "v1alpha1"
     static final String V1ALPHA3_API_VERSION = "v1alpha3"
     static final String V1BETA1_API_VERSION = "v1beta1"
     static final String APPLICATIONS = "applications"
@@ -31,6 +32,9 @@ class K8SClient {
     static final String SERVICE_CLASSES = "serviceclasses"
     static final String FUNCTIONS = "functions"
     static final String VIRTUAL_SERVICES = "virtualservices"
+
+    static final String TEST_LABEL = "test"
+    static final String TEST_LABEL_VALUE = "ac-e2e"
 
     private CustomObjectsApi customObjApi
 
@@ -57,15 +61,28 @@ class K8SClient {
         customObjApi.getNamespacedCustomObject(CONNECTOR_API_GROUP, V1ALPHA1_API_VERSION, namespace, APPLICATIONS, app)
     }
 
+    def deleteTestApplications() {
+        LinkedTreeMap response = customObjApi.listClusterCustomObject(CONNECTOR_API_GROUP, V1ALPHA1_API_VERSION, APPLICATIONS, "true", "test=ac-e2e", "", false)
+        List<Application> apps = (List<Application>) response.get("items")
+        apps.each { app ->
+            deleteApplicationMapping(app.metadata.name, KymaNames.INTEGRATION_NAMESPACE)
+            deleteApplication(app.metadata.name, KymaNames.INTEGRATION_NAMESPACE)
+        }
+    }
+
     def deleteApplication(String appName, String namespace) {
-        customObjApi.deleteNamespacedCustomObject(CONNECTOR_API_GROUP, V1ALPHA1_API_VERSION,
-                namespace,
-                APPLICATIONS,
-                appName,
-                new V1DeleteOptions(),
-                0,
-                null,
-                "Background")
+        try {
+            customObjApi.deleteNamespacedCustomObject(CONNECTOR_API_GROUP, V1ALPHA1_API_VERSION,
+                    namespace,
+                    APPLICATIONS,
+                    appName,
+                    new V1DeleteOptions(),
+                    0,
+                    null,
+                    "Background")
+        } catch (Exception e) {
+            println "Cannot delete the Application. Reason: " + e.toString()
+        }
     }
 
     def getTokenRequest(String appName, String namespace) {
@@ -75,7 +92,7 @@ class K8SClient {
     def createTokenRequest(String appName, String namespace) {
         TokenRequest tr = new TokenRequest().with {
             metadata = new Metadata(name: appName)
-            apiVersion = "${CONNECTOR_API_GROUP}/${v1ALPHA1_API_VERSION}"
+            apiVersion = "${CONNECTOR_API_GROUP}/${V1ALPHA1_API_VERSION}"
             kind = "TokenRequest"
             it
         }
@@ -84,39 +101,61 @@ class K8SClient {
     }
 
     def deleteTokenRequest(String appName, String namespace) {
-        customObjApi.deleteNamespacedCustomObject(CONNECTOR_API_GROUP, V1ALPHA1_API_VERSION, namespace,
-                TOKEN_REQUESTS,
-                appName,
-                new V1DeleteOptions(),
-                0,
-                null,
-                "Background")
+        try {
+            customObjApi.deleteNamespacedCustomObject(CONNECTOR_API_GROUP, V1ALPHA1_API_VERSION, namespace,
+                    TOKEN_REQUESTS,
+                    appName,
+                    new V1DeleteOptions(),
+                    0,
+                    null,
+                    "Background")
+        } catch (Exception e) {
+            println "Cannot delete the TokenRequest. Reason: " + e.toString()
+        }
     }
 
     def getApplicationMapping(String app, String namespace) {
         customObjApi.getNamespacedCustomObject(CONNECTOR_API_GROUP, V1ALPHA1_API_VERSION, namespace, APPLICATION_MAPPINGS, app)
     }
 
-    def createApplicationMapping(String appName, String namespace) {
-        ApplicationMapping am = new ApplicationMapping().with {
-            metadata = new Metadata(name: appName, namespace: namespace)
-            apiVersion = "${CONNECTOR_API_GROUP}/${v1ALPHA1_API_VERSION}"
-            kind = "ApplicationMapping"
-            it
-        }
+    def bindApplicationToNamespace(String appName, String namespace, boolean check = false) {
+        if (!check || !checkApplicationMappingExists(appName, namespace)) {
+            ApplicationMapping am = new ApplicationMapping().with {
+                metadata = new Metadata(name: appName, namespace: namespace)
+                apiVersion = "${CONNECTOR_API_GROUP}/${V1ALPHA1_API_VERSION}"
+                kind = "ApplicationMapping"
+                it
+            }
 
-        customObjApi.createNamespacedCustomObject(CONNECTOR_API_GROUP, V1ALPHA1_API_VERSION, namespace, APPLICATION_MAPPINGS, am, "true")
+            customObjApi.createNamespacedCustomObject(CONNECTOR_API_GROUP, V1ALPHA1_API_VERSION, namespace, APPLICATION_MAPPINGS, am, "true")
+        }
     }
 
     def deleteApplicationMapping(String appName, String namespace) {
-        customObjApi.deleteNamespacedCustomObject(CONNECTOR_API_GROUP, V1ALPHA1_API_VERSION,
-                namespace,
-                APPLICATION_MAPPINGS,
-                appName,
-                new V1DeleteOptions(),
-                0,
-                null,
-                "Background")
+        try {
+            if (checkApplicationMappingExists(appName, namespace)) {
+                customObjApi.deleteNamespacedCustomObject(CONNECTOR_API_GROUP, V1ALPHA1_API_VERSION,
+                        namespace,
+                        K8SClient.APPLICATION_MAPPINGS,
+                        appName,
+                        new V1DeleteOptions(),
+                        0,
+                        null,
+                        "Background")
+            }
+        } catch (Exception e) {
+            println "Cannot delete the ApplicationMapping. Reason: " + e.toString()
+        }
+    }
+
+    private def checkApplicationMappingExists(String app, String namespace) {
+        try {
+            getApplicationMapping(app, namespace)
+            return true
+        }
+        catch (final Exception ignore) {
+            return false
+        }
     }
 
     def getSubscription(String name, String namespace) {
@@ -129,14 +168,20 @@ class K8SClient {
     }
 
     def deleteSubscription(String subscriptionName, String namespace) {
-        customObjApi.deleteNamespacedCustomObject(EVENTING_API_GROUP, V1ALPHA1_API_VERSION,
-                namespace,
-                SUBSCRIPTIONS,
-                subscriptionName,
-                new V1DeleteOptions(),
-                0,
-                null,
-                "Background")
+        try {
+            customObjApi.deleteNamespacedCustomObject(EVENTING_API_GROUP, V1ALPHA1_API_VERSION,
+                    namespace,
+                    SUBSCRIPTIONS,
+                    subscriptionName,
+                    new V1DeleteOptions(),
+                    0,
+                    null,
+                    "Background")
+
+        }
+        catch (Exception e) {
+            println "Cannot delete the Subscription. Reason: " + e.toString()
+        }
     }
 
     def getServiceInstance(String name, String namespace) {
@@ -144,8 +189,25 @@ class K8SClient {
     }
 
     def createServiceInstance(ServiceInstance serviceInstance) {
-        def namespace = serviceInstance.metadata.namespace
-        customObjApi.createNamespacedCustomObject(SERVICE_CATALOG_API_GROUP, V1BETA1_API_VERSION, namespace, SERVICE_INSTANCES, serviceInstance, "true")
+        try {
+            def namespace = serviceInstance.metadata.namespace
+            customObjApi.createNamespacedCustomObject(SERVICE_CATALOG_API_GROUP, V1BETA1_API_VERSION, namespace, SERVICE_INSTANCES, serviceInstance, "true")
+        }
+        catch (Exception ignore) {
+            println "Cannot create the ServiceInstance."
+        }
+    }
+
+    def deleteServiceInstance(String name, String namespace) {
+        try {
+            customObjApi.deleteNamespacedCustomObject(SERVICE_CATALOG_API_GROUP, V1BETA1_API_VERSION, namespace, SERVICE_INSTANCES, name, new V1DeleteOptions(),
+                    0,
+                    null,
+                    "Background")
+        }
+        catch (final Exception e) {
+            println "Cannot delete the ServiceInstance. Reason: " + e.toString()
+        }
     }
 
     def getServiceClass(String serviceID, String namespace) {
@@ -158,14 +220,19 @@ class K8SClient {
     }
 
     def deleteLambdaFunction(String name, String namespace) {
-        customObjApi.deleteNamespacedCustomObject(KUBELESS_API_GROUP, V1BETA1_API_VERSION,
-                namespace,
-                FUNCTIONS,
-                name,
-                new V1DeleteOptions(),
-                0,
-                null,
-                "Background")
+        try {
+            customObjApi.deleteNamespacedCustomObject(KUBELESS_API_GROUP, V1BETA1_API_VERSION,
+                    namespace,
+                    FUNCTIONS,
+                    name,
+                    new V1DeleteOptions(),
+                    0,
+                    null,
+                    "Background")
+        }
+        catch (Exception e) {
+            println "Cannot delete the test Lambda. Reason: " + e.toString()
+        }
     }
 
     def getVirtualService(String name, String namespace) {
@@ -177,14 +244,18 @@ class K8SClient {
     }
 
     def deleteVirtualService(String name, String namespace) {
-        customObjApi.deleteNamespacedCustomObject(ISTIO_API_GROUP, V1ALPHA3_API_VERSION,
-                namespace,
-                VIRTUAL_SERVICES,
-                name,
-                new V1DeleteOptions(),
-                0,
-                null,
-                "Background")
+        try {
+            customObjApi.deleteNamespacedCustomObject(ISTIO_API_GROUP, V1ALPHA3_API_VERSION,
+                    namespace,
+                    VIRTUAL_SERVICES,
+                    name,
+                    new V1DeleteOptions(),
+                    0,
+                    null,
+                    "Background")
+        } catch (Exception e) {
+            println "Cannot delete the VirtualService. Reason: " + e.toString()
+        }
     }
 
     def getK8SService(String name, String namespace) {
@@ -197,13 +268,17 @@ class K8SClient {
     }
 
     void deleteService(String name, String namespace) {
-        coreApi.deleteNamespacedService(name,
-                namespace,
-                new V1DeleteOptions(),
-                "true",
-                0,
-                null,
-                "Background")
+        try {
+            coreApi.deleteNamespacedService(name,
+                    namespace,
+                    new V1DeleteOptions(),
+                    "true",
+                    0,
+                    null,
+                    "Background")
+        } catch (Exception e) {
+            println "Cannot delete the Service. Reason: " + e.toString()
+        }
     }
 
     def createDeployment(V1Deployment deployment) {
@@ -212,17 +287,21 @@ class K8SClient {
     }
 
     def deleteDeployment(String name, String namespace) {
-        appsApi.deleteNamespacedDeployment(name,
-                namespace,
-                new V1DeleteOptions(),
-                "true",
-                0,
-                null,
-                "Background")
+        try {
+            appsApi.deleteNamespacedDeployment(name,
+                    namespace,
+                    new V1DeleteOptions(),
+                    "true",
+                    0,
+                    null,
+                    "Background")
+        } catch (Exception e) {
+            println "Cannot delete the Service. Reason: " + e.toString()
+        }
     }
 
     def getPods(String name, String namespace) {
         coreApi.listNamespacedPod(namespace, "true", null, null, true, null, null, null, 30, false)
-                .items.stream().filter({pod -> pod.metadata.name.contains(name)}).collect(Collectors.toList())
+                .items.stream().filter({ pod -> pod.metadata.name.contains(name) }).collect(Collectors.toList())
     }
- }
+}
